feat(search): add NFT search filter and new components
