+++ frontend/src/components/StatsSection.jsx
@@ -1,0 +1,40 @@
+import { motion } from "framer-motion";
+
+const StatsSection = ({ stats }) => {
+  return (
+    <section className="bg-muted py-12">
+      <div className="container mx-auto px-4">
+        <motion.div
+          className="grid grid-cols-2 md:grid-cols-4 gap-8 text-center"
+          initial="hidden"
+          whileInView="visible"
+          viewport={{ once: true, amount: 0.3 }}
+          variants={{
+            hidden: { opacity: 0, y: 50 },
+            visible: {
+              opacity: 1,
+              y: 0,
+              transition: { duration: 0.8, staggerChildren: 0.2 },
+            },
+          }}
+        >
+          {stats.map((stat, index) => (
+            <motion.div
+              key={index}
+              className="space-y-2"
+              variants={{
+                hidden: { opacity: 0, y: 30 },
+                visible: { opacity: 1, y: 0 },
+              }}
+            >
+              <p className="text-3xl md:text-4xl font-bold">{stat.value}</p>
+              <p className="text-muted-foreground">{stat.label}</p>
+            </motion.div>
+          ))}
+        </motion.div>
+      </div>
+    </section>
+  );
+};
+
+export default StatsSection;
